GEODE-9997: move API impacts to InternalGatewaySender
