+++ geode-core/src/main/java/org/apache/geode/internal/cache/GemFireCacheImpl.java
@@ -233,6 +233,7 @@ import org.apache.geode.internal.cache.tier.sockets.ServerConnection;
 import org.apache.geode.internal.cache.wan.AbstractGatewaySender;
 import org.apache.geode.internal.cache.wan.GatewaySenderAdvisor;
 import org.apache.geode.internal.cache.wan.GatewaySenderQueueEntrySynchronizationListener;
+import org.apache.geode.internal.cache.wan.InternalGatewaySender;
 import org.apache.geode.internal.cache.wan.WANServiceProvider;
 import org.apache.geode.internal.cache.wan.parallel.ParallelGatewaySenderQueue;
 import org.apache.geode.internal.cache.xmlcache.CacheServerCreation;
@@ -2186,7 +2187,7 @@ public class GemFireCacheImpl implements InternalCache, InternalClientCache, Has
 
       for (GatewaySender sender : allGatewaySenders) {
         try {
-          sender.prepareForStop();
+          ((InternalGatewaySender) sender).prepareForStop();
         } catch (Exception exception) {
           if (isDebugEnabled) {
             logger.debug("When calling Prepare for stop gw sender, ignore exception " + exception);
