+++ geode-core/src/main/java/org/apache/geode/internal/cache/wan/InternalGatewaySender.java
@@ -49,4 +49,8 @@ public interface InternalGatewaySender extends GatewaySender {
   void setStartEventProcessorInPausedState();
 
   int getEventQueueSize();
+
+
+  void prepareForStop();
+
 }
