+++ geode-wan/src/distributedTest/java/org/apache/geode/internal/cache/wan/parallel/ParallelGatewaySenderOperationsDistributedTest.java
@@ -1262,7 +1262,6 @@ public class ParallelGatewaySenderOperationsDistributedTest extends WANTestBase 
 
   }
 
-
   private void clearShadowBucketRegions(PartitionedRegion shadowRegion) {
     PartitionedRegionDataStore.BucketVisitor bucketVisitor =
         (bucketId, r) -> ((BucketRegion) r).clearEntries(null);
